requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
faker==20.1.0
APScheduler==3.10.4
alembic==1.12.1
openpyxl==3.1.2
psycopg2-binary==2.9.9
aiosmtplib==3.0.1
email-validator==2.1.0